# Make UnsafeConversion fast by inlining copies