--- a/pkg/api/conversion.go
+++ b/pkg/api/conversion.go
@@ -158,7 +158,6 @@ func Convert_fields_Selector_To_string(in *fields.Selector, out *string, s conve
 	return nil
 }
 func Convert_resource_Quantity_To_resource_Quantity(in *resource.Quantity, out *resource.Quantity, s conversion.Scope) error {
-	// Cannot deep copy these, because inf.Dec has unexported fields.
-	*out = *in.Copy()
+	*out = *in
 	return nil
 }