… pkg/api/conversion_test.go
… pkg/api/v1/conversion_generated.go
… pkg/apis/batch/v1/conversion_generated.go
… pkg/apis/batch/v2alpha1/conversion_generated.go
… pkg/apis/extensions/v1beta1/conversion_generated.go
… pkg/apis/policy/v1alpha1/conversion_generated.go
… pkg/apis/rbac/v1alpha1/conversion_generated.go
… pkg/util/diff/diff.go
… pkg/util/diff/diff_test.go