@smarterclayton 2016-05-21
Ok, this should be ready for review.  Basically, any field that is convertible directly to the output field is copied over.  Any conversion that mutates the object must be refactored.  Added a guard to the serialization test that validates that the input object is not mutated.  Made ConvertToVersion "safe" by adding a copy.  Added a reflection based object diff (that I will switch ObjectDiff over to using in the future) to make debugging deltas not a PITA.