@smarterclayton 2016-05-21
```
$ benchcmp /tmp/old /tmp/new
benchmark                                      old ns/op     new ns/op     delta
BenchmarkPodConversion-8                       5650          4761          -15.73%
BenchmarkNodeConversion-8                      5450          4817          -11.61%
BenchmarkReplicationControllerConversion-8     6817          4090          -40.00%
BenchmarkEncodeCodecFromInternalProtobuf-8     5035          4765          -5.36%
BenchmarkDecodeCodecToInternalProtobuf-8       6426          6244          -2.83%
BenchmarkEncodeCodecFromInternal-8             20761         20599         -0.78%
BenchmarkDecodeCodec-8                         48350         46996         -2.80%
BenchmarkDecodeIntoInternalCodec-8             48580         52025         +7.09%

benchmark                                      old allocs     new allocs     delta
BenchmarkPodConversion-8                       36             24             -33.33%
BenchmarkNodeConversion-8                      26             18             -30.77%
BenchmarkReplicationControllerConversion-8     35             20             -42.86%
BenchmarkEncodeCodecFromInternalProtobuf-8     17             13             -23.53%
BenchmarkDecodeCodecToInternalProtobuf-8       57             53             -7.02%
BenchmarkEncodeCodecFromInternal-8             48             43             -10.42%
BenchmarkDecodeCodec-8                         195            191            -2.05%
BenchmarkDecodeIntoInternalCodec-8             197            193            -2.03%
```