@smarterclayton 2016-05-21
6-8us for a decode is pretty good.  I'll call that a win for 1.3.