@smarterclayton 2016-05-21
After this change, there really isn't much fat left in conversion and decode.  Probably some odds and ends.