@wojtek-t 2016-05-21
@smarterclayton - I took a very brief look into it and added few minor comments in individual commits. Let me do the final review on Monday morning.