@wojtek-t 2016-05-23
@smarterclayton - I added few more comments in the individual commits.