@smarterclayton 2016-05-23
@k8s-bot test this issue #IGNORE (GitHub is going up and down today)